Testing Wavelet Tree implementation

+++ src/bitvector.cpp
@@ -194,8 +194,8 @@ void bitVector::append1() {
     set1(_size++);
 }
 
-unsigned long bitVector::size() const { return _size; }
-unsigned long bitVector::cap() const { return _cap; }
+size_t bitVector::size() const { return _size; }
+size_t bitVector::cap() const { return _cap; }
 
 /**
     Coloca uma sequencia predefinida de bits ao final do bitvector.
